requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
streamlit-chat>=0.1.1
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4

# RAG and AI Components
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1

# Vector Storage and Embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2
huggingface-hub>=0.17.0

# Document Processing
pypdf>=3.17.0
python-docx>=0.8.11
python-pptx>=0.6.21
beautifulsoup4>=4.12.2
requests>=2.31.0

# Language Processing and Detection
langdetect>=1.0.9
googletrans>=4.0.0rc1
polyglot>=16.7.4
textblob>=0.17.1

# Voice Processing
speech-recognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
gTTS>=2.4.0

# Database and Authentication
supabase>=1.0.4
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.21

# Data Processing and Analytics
pandas>=2.1.0
numpy>=1.24.3
plotly>=5.17.0
matplotlib>=3.7.2
seaborn>=0.12.2

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3

# HTTP and API
httpx>=0.25.0
fastapi>=0.103.0
uvicorn>=0.23.2

# Utilities
tqdm>=4.66.0
rich>=13.5.2
click>=8.1.7
python-multipart>=0.0.6

# Development and Testing
pytest>=7.4.2
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.1

# Security
cryptography>=41.0.4
passlib>=1.7.4
python-jose>=3.3.0

# File Handling
aiofiles>=23.2.1
python-magic>=0.4.27

# Caching and Performance
redis>=5.0.0
diskcache>=5.6.3

# Monitoring and Logging
loguru>=0.7.0
sentry-sdk>=1.32.0
